% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_type.R
\name{plot_type}
\alias{plot_type}
\title{Visualize the genres types}
\usage{
plot_type(data)
}
\arguments{
\item{data}{Object to be plotted}
}
\description{
Visualize the genres types
}
\author{
Chantel Wetzel
}
